specify dtype

+++ src/daria/corrections/shape/curvature.py
@@ -455,8 +455,8 @@ class CurvatureCorrection:
         Ny, Nx = img.shape[:2]
 
         # Define coordinates
-        x = np.arange(Nx)
-        y = np.arange(Ny)
+        x = np.arange(Nx, dtype=np.int64)
+        y = np.arange(Ny, dtype=np.int64)
 
         print("dtype of coordinates", x.dtype)
